feat: add explicit support for npm/yarn/pnpm/bun workspaces
Refactored workspace detection to separate generic workspace support from Turbo-specific logic:

- Created new `workspaces.rs` module for generic npm/yarn/pnpm/bun workspace detection
- Simplified `turbo.rs` to focus only on turbo.json detection, delegating to workspaces module
- Updated discovery chain in `mod.rs` to check: Nx → Turbo → generic workspaces
- Updated README to reflect support for npm/yarn/pnpm/bun workspaces

This makes workspace support more explicit and maintainable while preserving backward compatibility.
